Extract the file-sync filename grammar into fileSyncNames
Pure code motion, first of the incremental fileSyncConnection.ts extractions: the nine pure name predicates and constructors (parseMessageByteCount, parseTimestampedMessageNNN, ackMarkerName, peerIdFromControlName, isProtocolTempName, isAbortMarkerName, isExpectedAbortName, isProtocolGrammarName, isRetainMessageAck) and the four grammar suffix constants move to packages/core/src/connection/fileSyncNames.ts. fileSyncConnection.ts imports them back and re-exports isAbortMarkerName and isExpectedAbortName so the public surface through main.ts is unchanged. Helpers that close over instance state (messageFilename, isPeerHelloName, isRecognizedLoopFile, and friends) deliberately stay behind; they belong to the state-machine split.
